requirements.txt
matplotlib==3.5.2
numpy==1.22.2
pandas==1.4.1
Pillow==9.2.0
PyYAML==6.0
torch==1.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.63.0
